+++ yak_runner/lib/src/delegates/all.dart
@@ -1,3 +1,2 @@
-export 'error/all.dart';
 export 'function/all.dart';
 export 'test/all.dart';