+++ yak_runner/lib/src/mixin/result/on_result/result.dart
@@ -1,5 +1,5 @@
 import 'package:yak_runner/src/classes/failure/all.dart';
-
+import '../../../typedef/all.dart';
 import '../../../classes/all.dart';
 
 mixin OnResult<S> on ResultBase<S> {