+++ yak_runner/lib/src/runners/async/unary.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
-import 'package:yak_runner/src/shared/all.dart';
-
+import '../../shared/all.dart';
+import '../../typedef/all.dart';
 import '../../classes/all.dart';
 import '../../mixin/all.dart';
 import '../../delegates/all.dart';
@@ -13,7 +13,7 @@ class YakRunnerArgAsync<T, S>
       : assert(fun != null, ' a non null function must be provided'),
         assert(S != typeVoid, '`S` must not be of type `void`');
   final Future<T> Function(FutureOr<S>) fun;
-  final void Function(Object, StackTrace) errorHandler;
+  final Catch errorHandler;
 
   Future<Result<T>> call(FutureOr<S> arg) async {
     assert(arg != null, ' a non null `arg` must be provided');