+++ yak_error_handler/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.0+1] - 14 January 2021
+
+* added example, updated readme &
+
 ## [0.1.0+0] - initila release
 
 * initial implementation complete, all tests passing